# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SGMA (Sistema de Gestión de Mantenimiento y Activos)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [work-orders]
      summary: Listar órdenes de trabajo
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            type: string
        - name: tipo
          in: query
          required: false
          schema:
            type: string
        - name: prioridad
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [work-orders]
      summary: Crear orden de trabajo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /work-orders/{id}:
    get:
      operationId: getWorkOrder
      tags: [work-orders]
      summary: Obtener orden de trabajo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateWorkOrder
      tags: [work-orders]
      summary: Actualizar orden de trabajo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    delete:
      operationId: deleteWorkOrder
      tags: [work-orders]
      summary: Eliminar orden de trabajo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    WorkOrder:
      type: object
      required:
        - id
        - numero
        - equipoId
        - tipo
        - prioridad
        - estado
        - fechaCreacion
      properties:
        id:
          type: integer
        numero:
          type: string
        equipoId:
          type: integer
        equipoNombre:
          type: ["string", "null"]
        equipoCodigo:
          type: ["string", "null"]
        tipo:
          type: string
        prioridad:
          type: string
        estado:
          type: string
        tecnicoId:
          type: ["integer", "null"]
        tecnicoNombre:
          type: ["string", "null"]
        centroTrabajoId:
          type: ["integer", "null"]
        centroTrabajoNombre:
          type: ["string", "null"]
        descripcion:
          type: ["string", "null"]
        reporteFalla:
          type: ["string", "null"]
        diagnostico:
          type: ["string", "null"]
        causaRaiz:
          type: ["string", "null"]
        solucion:
          type: ["string", "null"]
        horasEstimadas:
          type: ["number", "null"]
        horasReales:
          type: ["number", "null"]
        costoManoObra:
          type: ["number", "null"]
        costoRepuestos:
          type: ["number", "null"]
        costoTotal:
          type: ["number", "null"]
        fechaCreacion:
          type: string
        fechaProgramada:
          type: ["string", "null"]
        fechaCierre:
          type: ["string", "null"]
    WorkOrderInput:
      type: object
      required:
        - equipoId
        - tipo
        - prioridad
        - estado
      properties:
        equipoId:
          type: integer
        tipo:
          type: string
        prioridad:
          type: string
        estado:
          type: string
        tecnicoId:
          type: integer
        centroTrabajoId:
          type: integer
        descripcion:
          type: string
        reporteFalla:
          type: string
        diagnostico:
          type: string
        causaRaiz:
          type: string
        solucion:
          type: string
        horasEstimadas:
          type: number
        horasReales:
          type: number
        costoManoObra:
          type: number
        costoRepuestos:
          type: number
        fechaProgramada:
          type: string
    WorkOrderUpdate:
      type: object
      properties:
        equipoId:
          type: integer
        tipo:
          type: string
        prioridad:
          type: string
        estado:
          type: string
        tecnicoId:
          type: integer
        centroTrabajoId:
          type: integer
        descripcion:
          type: string
        reporteFalla:
          type: string
        diagnostico:
          type: string
        causaRaiz:
          type: string
        solucion:
          type: string
        horasEstimadas:
          type: number
        horasReales:
          type: number
        costoManoObra:
          type: number
        costoRepuestos:
          type: number
        fechaProgramada:
          type: string
        fechaCierre:
          type: string
...